Hotfix for GOG Header change + minor changes
GOG Headers sometimes return "last-modified" instead of "Last-Modified" to disastrous effect.  Fixed by checking for the lower case version if the proper one fails.

Minor changes for compatibility:
Removed an expression shortcut that does't work in Python 2.
Added 'md5_xmls' to the Orphan Exclusion List ( should help for non-standard directory layout where you're downloading to the folder containing gogrepoc.py )
Added notification that Google/Discord logins are currently unsupported ( is on the to do list but not high priority, as I have had 0 requests for this)